feat: useDeviceWidth 추가 후 store page 적용
